1.62
- SA: Shift frequency bands slightly to better match mic frequency  response and possibly batch-to-batch dispersion; better pace-keeping; clean up arduinoFFT.
- IO14 is now mirror of the TCD's TT-OUT (in its original function); goes HIGH  on start of temporal displacement, LOW on re-entry. There is no lead time. Note: There is no terminal for IO14. It is, however, accessible through a solder pad next to the ESP32, pin no. 8 from bottom.
- BTTFN enhancements in sync with TCD capabilities
- Code refinements

+++ sid-A10001986/input.cpp
@@ -1,7 +1,7 @@
 /*
  * -------------------------------------------------------------------
  * CircuitSetup.us Status Indicator Display
- * (C) 2023-2025 Thomas Winischhofer (A10001986)
+ * (C) 2023-2026 Thomas Winischhofer (A10001986)
  * https://github.com/realA10001986/SID
  * https://sid.out-a-ti.me
  *
@@ -157,7 +157,7 @@ bool IRRemote::loop()
 
     // Copy result to backup buffer
     _buflen = _irlen;
-    for(uint8_t i = 0; i < _buflen; i++) {
+    for(int i = 0; i < _buflen; i++) {
         _buf[i] = _irbuf[i];
     }
